Closes #117
