+++ Tribler/Test/TUPT/test_TUPTControl.py
@@ -7,7 +7,7 @@ from Tribler.TUPT.TorrentFinder.TorrentFinderControl import TorrentFinderControl
 from Tribler.TUPT.Movie import Movie
 
 from Tribler.Test.TUPT.test_StubPluginManager import PluginManagerStub
-from Tribler.Test.TUPT.test_StubPluginManager import TorrentDefStub
+from Tribler.Test.TUPT.TorrentFinder.TorrentFinderStubs import TorrentDefStub
   
 
 class TestMovieTorrentIterator(unittest.TestCase):
